refactor: move destination mutation to dedicated file
